+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+### Changes
+
+* `JsonConfig` is now `Send + Sync`, this implies that `error_handler` must be `Send + Sync` too.
+
 ## [1.0.0-beta.4] - 2019-05-12
 
 ### Add
@@ -9,9 +13,7 @@
 ### Changes
 
 * `App::configure` take an `FnOnce` instead of `Fn`
-
 * Upgrade actix-net crates
-
 
 ## [1.0.0-beta.3] - 2019-05-04