+++ CHANGES.md
@@ -1,5 +1,24 @@
 # Changes
 
+### Add
+
+* `QueryConfig`, similar to `JsonConfig` for customizing error handling of query extractors.
+
+### Changes
+
+* `JsonConfig` is now `Send + Sync`, this implies that `error_handler` must be `Send + Sync` too.
+
+## [1.0.0-beta.4] - 2019-05-12
+
+### Add
+
+* Allow to set/override app data on scope level
+
+### Changes
+
+* `App::configure` take an `FnOnce` instead of `Fn`
+* Upgrade actix-net crates
+
 ## [1.0.0-beta.3] - 2019-05-04
 
 ### Added